Rename namespace PublicationFieldValues to PublicationFields
