+++ packages/framework/src/Console/Commands/MakePublicationCommand.php
@@ -9,11 +9,11 @@ use Hyde\Console\Commands\Helpers\InputStreamHandler;
 use Hyde\Console\Concerns\ValidatingCommand;
 use Hyde\Framework\Actions\CreatesNewPublicationPage;
 use Hyde\Framework\Features\Publications\Models\PublicationFieldDefinition;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\ArrayField;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\ImageField;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\PublicationField;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\TagField;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\TextField;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\ArrayField;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\ImageField;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\PublicationField;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\TagField;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\TextField;
 use Hyde\Framework\Features\Publications\Models\PublicationType;
 use Hyde\Framework\Features\Publications\PublicationFieldTypes;
 use Hyde\Framework\Features\Publications\PublicationService;
