+++ packages/framework/src/Framework/Features/Publications/Models/PublicationFields/ArrayField.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Features\Publications\Models\PublicationFieldValues;
+namespace Hyde\Framework\Features\Publications\Models\PublicationFields;
 
 use Hyde\Framework\Features\Publications\PublicationFieldTypes;
 use function is_array;
 
-final class TagField extends PublicationField
+final class ArrayField extends PublicationField
 {
-    public const TYPE = PublicationFieldTypes::Tag;
+    public const TYPE = PublicationFieldTypes::Array;
 
     public function __construct(string|array $value = null)
     {
