+++ packages/framework/src/Framework/Features/Publications/Models/PublicationFields/Concerns/CanonicableTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Features\Publications\Models\PublicationFieldValues\Concerns;
+namespace Hyde\Framework\Features\Publications\Models\PublicationFields\Concerns;
 
 use RuntimeException;
 use function substr;
