+++ packages/framework/src/Framework/Features/Publications/Models/PublicationFields/Contracts/Canonicable.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Features\Publications\Models\PublicationFieldValues\Contracts;
+namespace Hyde\Framework\Features\Publications\Models\PublicationFields\Contracts;
 
 use Stringable;
