+++ packages/framework/src/Framework/Features/Publications/Models/PublicationFields/IntegerField.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Features\Publications\Models\PublicationFieldValues;
+namespace Hyde\Framework\Features\Publications\Models\PublicationFields;
 
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\Concerns\CanonicableTrait;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\Contracts\Canonicable;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\Concerns\CanonicableTrait;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\Contracts\Canonicable;
 use Hyde\Framework\Features\Publications\PublicationFieldTypes;
 
 final class IntegerField extends PublicationField implements Canonicable
