+++ packages/framework/tests/Feature/Actions/CreatesNewPublicationPageTest.php
@@ -7,9 +7,9 @@ namespace Hyde\Framework\Testing\Feature\Actions;
 use function file_get_contents;
 use Hyde\Facades\Filesystem;
 use Hyde\Framework\Actions\CreatesNewPublicationPage;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\StringField;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\TagField;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\TextField;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\StringField;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\TagField;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\TextField;
 use Hyde\Framework\Features\Publications\Models\PublicationType;
 use Hyde\Hyde;
 use Hyde\Testing\TestCase;
