+++ packages/framework/tests/Features/Publications/Models/PublicationFieldValues/Concerns/CanonicableTraitTest.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Features\Publications\Models\PublicationFieldValues\Concerns;
 
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\Concerns\CanonicableTrait;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\Contracts\Canonicable;
-use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\PublicationField;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\Concerns\CanonicableTrait;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\Contracts\Canonicable;
+use Hyde\Framework\Features\Publications\Models\PublicationFields\PublicationField;
 use Hyde\Testing\TestCase;
 use RuntimeException;
 
 /**
- * @covers \Hyde\Framework\Features\Publications\Models\PublicationFieldValues\Concerns\CanonicableTrait
+ * @covers \Hyde\Framework\Features\Publications\Models\PublicationFields\Concerns\CanonicableTrait
  */
 class CanonicableTraitTest extends TestCase
 {
